[Fleet] Remove server code from common folder in ingest
