+++ x-pack/legacy/plugins/ingest/server/libs/datasources.contract.test.ts
@@ -10,7 +10,8 @@ import { INDEX_NAMES } from '../../common/constants/index_names';
 import { FrameworkUser } from './adapters/framework/adapter_types';
 import { exampleStoredDatasource } from './adapters/datasource/adapter_types';
 import { compose } from './compose/memorized';
-import { ServerLibs, Datasource } from './types';
+import { ServerLibs } from './types';
+import { Datasource } from '../../common/types/domain_data';
 
 jest.mock('uuid/v4', () => {
   let uuid = 1;
